% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_category}
\alias{food_category}
\title{Food categories}
\format{
A data frame with 28 rows and 3 columns:
\describe{
\item{id}{Unique permanent identifier.}
\item{code}{Food group code.}
\item{description}{Description of the food group.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
food_category
}
\description{
Foods of defined similarity.
}
\keyword{datasets}
