% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_nutrient}
\alias{food_nutrient}
\title{Food nutrients}
\format{
A data frame with 19,828 rows and 11 columns:
\describe{
\item{id}{Unique permanent identifier.}
\item{fdc_id}{ID of the food this food nutrient pertains to.
Foreign key to \link{food}.}
\item{nutrient_id}{ID of the nutrient to which the food nutrient
pertains. Foreign key to \link{nutrient}.}
\item{amount}{Amount of the nutrient per 100g of food.}
\item{data_points}{Number of observations on which the value is
based.}
\item{derivation_id}{ID of the food nutrient derivation technique
used to derive the value.}
\item{min}{The minimum amount.}
\item{max}{The maximum amount.}
\item{median}{The median amount.}
\item{footnote}{Comments on any unusual aspects of the food
nutrient.}
\item{min_year_acquired}{Minimum purchase year of all acquisitions
used to derive the nutrient value.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
food_nutrient
}
\description{
A nutrient value for a food. Amounts are per 100g of food, in the
unit defined in the \link{nutrient} table.
}
\keyword{datasets}
