% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_portion}
\alias{food_portion}
\title{Food portions}
\format{
A data frame with 187 rows and 11 columns:
\describe{
\item{id}{Unique permanent identifier.}
\item{fdc_id}{ID of the food this food portion pertains to.
Foreign key to \link{food}.}
\item{seq_num}{The order the measure will be displayed on the
released food.}
\item{amount}{The number of measure units that comprise the
measure (e.g. if measure is 3 tsp, the amount is 3).}
\item{measure_unit_id}{The unit used for the measure. Foreign key
to \link{measure_unit}.}
\item{portion_description}{Comments that provide more specificity
on the measure.}
\item{modifier}{Qualifier of the measure (e.g. related to food
shape or form).}
\item{gram_weight}{The weight of the measure in grams.}
\item{data_points}{The number of observations on which the
measure is based.}
\item{footnote}{Comments on any unusual aspects of the measure.}
\item{min_year_acquired}{Minimum purchase year of all acquisitions
used to derive the measure value.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
food_portion
}
\description{
Discrete amounts of food, used to convert from per-100g nutrient
values to common serving sizes.
}
\keyword{datasets}
