% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{foodbank_sql}
\alias{foodbank_sql}
\title{Generate SQL to create DuckDB views for all foodbank tables}
\usage{
foodbank_sql()
}
\value{
A character vector of SQL \verb{CREATE VIEW} statements.
}
\description{
Returns a vector of SQL statements that create DuckDB views backed
by the parquet files included in the package.
}
\examples{
foodbank_sql()

if (requireNamespace("duckdb", quietly = TRUE)) {
  con <- DBI::dbConnect(duckdb::duckdb())
  for (sql in foodbank_sql()) DBI::dbExecute(con, sql)
  DBI::dbListTables(con)
  DBI::dbDisconnect(con)
}
}
