% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_unit}
\alias{measure_unit}
\title{Measure units}
\format{
A data frame with 123 rows and 2 columns:
\describe{
\item{id}{Unique permanent identifier.}
\item{name}{Name of the unit.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
measure_unit
}
\description{
Units for measuring quantities of foods.
}
\keyword{datasets}
