% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nutrient}
\alias{nutrient}
\title{Nutrients}
\format{
A data frame with 477 rows and 5 columns:
\describe{
\item{id}{Unique permanent identifier.}
\item{name}{Name of the nutrient.}
\item{unit_name}{The standard unit of measure for the nutrient
(per 100g of food).}
\item{nutrient_nbr}{A unique code identifying a nutrient or food
constituent.}
\item{rank}{Display rank for ordering nutrients.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
nutrient
}
\description{
The chemical constituents of a food (e.g. calcium, vitamin E)
officially recognized as essential to human health.
}
\keyword{datasets}
