% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food}
\alias{food}
\title{Foods}
\format{
A data frame with 436 rows and 5 columns:
\describe{
\item{fdc_id}{Unique permanent identifier of the food.}
\item{data_type}{Type of food data.}
\item{description}{Description of the food.}
\item{food_category_id}{ID of the food category the food belongs
to. Foreign key to \link{food_category}.}
\item{publication_date}{Date when the food was published to
FoodData Central.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
food
}
\description{
Any substance consumed by humans for nutrition, taste and/or aroma.
From the USDA FoodData Central Foundation Foods dataset
(December 2025).
}
\keyword{datasets}
