% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measure_unit}
\alias{measure_unit}
\title{Measure units}
\format{
A data frame with 123 rows and 2 columns:
\describe{
\item{id}{Unique permanent identifier.}
\item{name}{Name of the unit.}
}
}
\source{
\url{https://fdc.nal.usda.gov/download-datasets}
}
\usage{
measure_unit
}
\description{
Units for measuring quantities of foods.
}
\section{Joins}{

\itemize{
\item \link{food_portion} on \code{id = measure_unit_id}: portions using this
unit.
}
}

\examples{
\dontshow{if (requireNamespace("dplyr", quietly = TRUE)) withAutoprint(\{ # examplesIf}
library(dplyr)

# See which units are used in portions
measure_unit |>
  inner_join(food_portion, join_by(id == measure_unit_id))
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
